% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_draw_geometry.R
\name{ee_draw_geometry}
\alias{ee_draw_geometry}
\title{Draw object on Earth Engine}
\usage{
ee_draw_geometry(
  name,
  import_as = "A",
  properties = NULL,
  color = "red",
  zoom = 5,
  ...
)
}
\arguments{
\item{name}{name of new object on R.}

\item{import_as}{type of Earth Engine object.}

\item{properties}{is a list with properties of EE object.}

\item{color}{is a color of geometry (hexadecimal).}

\item{zoom}{is the zoom of the visualization in the display.}

\item{...}{optional arguments for Map object.}
}
\value{
An Earth Engine object, depends of the import_as, if is of type:
\itemize{
\item \bold{"A":} return a geometry.
\item \bold{"B":} return a feature.
\item \bold{"C":} return a feature collection.
}
}
\description{
Function to draw an Earth Engine object interactively from R
}
\details{
import_as
\itemize{
\item \bold{"A":} An Earth Engine object of type Geometry.
\item \bold{"B":} An Earth Engine object of type Feature.
\item \bold{"C":} An Earth Engine object of type Feature collection.
}
}
